Whitelist and blacklist modify compiled blocklist directly
This means that you do not need to call `pihole.compile_list()`
after adding or removing a domain from the white/blacklist.
If you want a freshly compiled ad list, you can still compile
it, but the process takes much longer because it has to remake
the ad list entirely. This method skips that long wait by only
making the necessary changes.