Remove error code only output
User-facing scripts shouldn't need to worry about outputting in a
way for other scripts to understand. For example: Flask will use
the Pihole class's whitelist methods instead of the whitelist
script, because it can get the data it needs from the return values
and possible thrown exceptions of those methods.